requirements.txt
alembic==1.12.1
autopep8==2.0.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
exceptiongroup==1.2.1
Flask==2.2.5
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
httpie==3.2.2
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
packaging==24.0
pluggy==1.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.10.0
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.4
pytest-flask==1.3.0
pytest-mock==3.11.1
python-dotenv==0.21.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
